tests/wpt/web-platform-tests/webmessaging/with-options/unknown-parameter.tentative.html

<!doctype html>
<title>unknown parameter</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  postMessage('', [], {someBogusParameterOnThisDictionary: 'food'});
  onmessage = this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  });
});
</script>
